requirements.txt
strands-agents
strands-agents-tools
boto3
twilio  # optional — only needed for WhatsApp alerts
